+++ src/defaults.js
@@ -1,7 +1,7 @@
 module.exports = {
     handleQuota: true,
     QPS: 2,
-    forwarderAddress: "app/integ/forwarder/1.3/apiForwarder.html",
+    forwarderAddress: "app/integ/forwarder/1.4/apiForwarder.html",
     channelMarker: "'E",
     httpRequest: null,
     oldIEForwarder: false